Komunikacja samochodów correct version

+++ src/main/java/SmartCity/Crossroad.java
@@ -9,14 +9,14 @@ import org.jxmapviewer.viewer.Waypoint;
 import org.jxmapviewer.viewer.WaypointPainter;
 
 public abstract class Crossroad {
-	public abstract void addCarToQueue(String carName, int adjacentOsmWayId);
-	public abstract void addCarToFarAwayQueue(String carName, int adjacentOsmWayId, int journeyTime);
-	public abstract void removeCarFromQueue(int adjacentOsmWayId);
-	public abstract void removeCarFromFarAwayQueue(String carName, int adjacentOsmWayId);
-	public abstract void addPedestrianToQueue(String pedestrianName, int adjacentOsmWayId);
-	public abstract void removePedestrianFromQueue(String pedestrianName, int adjacentOsmWayId);
+	public abstract void addCarToQueue(String carName, long adjacentOsmWayId);
+	public abstract void addCarToFarAwayQueue(String carName, long adjacentOsmWayId, int journeyTime);
+	public abstract void removeCarFromQueue(long adjacentOsmWayId);
+	public abstract void removeCarFromFarAwayQueue(String carName, long adjacentOsmWayId);
+	public abstract void addPedestrianToQueue(String pedestrianName, long adjacentOsmWayId);
+	public abstract void removePedestrianFromQueue(String pedestrianName, long adjacentOsmWayId);
 	public abstract OptimizationResult requestOptimizations();
-	public abstract boolean isLightGreen(int adjacentOsmWayId);
+	public abstract boolean isLightGreen(long adjacentOsmWayId);
 	public abstract void draw( List<Painter<JXMapViewer>>painter);
 	public abstract void startLifetime();
 }
